DCGAN/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.3.0
Pillow>=8.0.0
tqdm>=4.60.0
